Add files via upload
This Register-Login app shows typical register-login procedure towards reaching a particular page(s) of interest. I am using a tutorial lesson as an example in this mini project. The logic can however be used for various register-login procedures in many applications.
The genKey.js file, which contains code to generate the secret JWT key in not included in the uploaded files.
I really look forward to comments.